refactor : bascket model-domain 분리
