+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/model/bascket/repository/BasketRepository.java
@@ -1,7 +1,7 @@
-package com.sparta.blackwhitedeliverydriver.bascket.repository;
+package com.sparta.blackwhitedeliverydriver.model.bascket.repository;
 
-import com.sparta.blackwhitedeliverydriver.bascket.entity.Basket;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.bascket.entity.Basket;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
